comment what functions do for better clarity

+++ philo/timing.c
@@ -3,34 +3,41 @@
 /*                                                        :::      ::::::::   */
 /*   timing.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lmedrano <marvin@42lausanne.ch>            +#+  +:+       +#+        */
+/*   By: lmedrano <lmedrano@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/11 12:41:17 by lmedrano          #+#    #+#             */
-/*   Updated: 2023/08/15 16:40:49 by lmedrano         ###   ########.fr       */
+/*   Updated: 2023/08/16 13:25:30 by lmedrano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	get_current_time(void)
+//FCT GET_CURRENT_TIME
+// Utilise gettimeofday pour obtenir le temps actuel
+// en millisecondes depuis une référence temporelle.
+// Renvoie le temps actuel en millisecondes.
+unsigned long long	get_current_time(void)
 {
-	struct timeval	time;
-	double			current_time;
+	struct timeval		time;
+	unsigned long long	current_time;
 
 	gettimeofday(&time, NULL);
 	current_time = (time.tv_sec * 1000) + (time.tv_usec / 1000);
 	return (current_time);
 }
 
-long int	time_passed(long int start, long int end)
+//FCT TIME_PASSED
+//permet de calculer la durée écoulée entre deux moments dans le temps.
+unsigned long long	time_passed(unsigned long long start,
+	unsigned long long end)
 {
 	return (end - start);
 }
 
-void	ft_usleep(long int time)
+void	ft_usleep(unsigned long long time)
 {
-	long int	sleep;
-	long int	current_time;
+	unsigned long long int	sleep;
+	unsigned long long int	current_time;
 
 	current_time = get_current_time();
 	sleep = time_passed(current_time, get_current_time());
